backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
dateparser==1.1.8
spacy==3.7.2
pydantic==2.4.2
email-validator==2.0.0
# Email dependencies are built into Python standard library 